# openapi.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: Documentation for the API of an app that handles with users and phone numbers of several organizations.

paths:
  /api/organizations:
    get:
      summary: Get Organizations
      description: Fetches a list of organizations.
      responses:
        '200':
          description: A list of organizations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '500':
          description: Error getting organizations.

  /api/phones:
    get:
      summary: Get Phones
      description: Fetches a list of phones based on the provided organization ID.
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The ID of the organization whose phones are to be retrieved.
      responses:
        '200':
          description: A list of phones.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Phone'
        '400':
          description: ID is required.
        '500':
          description: Error getting phones.

    /api/users:
      get:
        summary: Get Users
        description: Fetches a list of users based on the provided organization ID.
        parameters:
          - in: query
            name: id
            schema:
              type: string
            required: true
            description: The ID of the organization whose users are to be retrieved.
        responses:
          '200':
            description: A list of users.
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/User'
          '400':
            description: ID is required.
          '500':
            description: Error getting users.

      delete:
        summary: Delete User
        description: Deletes a user based on the provided ID.
        parameters:
          - in: query
            name: id
            schema:
              type: string
            required: true
            description: The ID of the user to delete.
        responses:
          '200':
            description: User deleted successfully.
            content:
              application/json:
                schema:
                  type: object
          '400':
            description: ID is required.
          '500':
            description: Error deleting user.

      post:
        summary: Create User
        description: Creates a new user.
        requestBody:
          description: User object to be created.
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        responses:
          '200':
            description: User created successfully.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
          '500':
            description: Error creating user.

    /api/phones:
      patch:
        summary: Update Phones
        description: Updates the phone details.
        requestBody:
          description: Phone object with updated details.
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Phone'
        responses:
          '200':
            description: Phone updated successfully.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                    type: boolean
                    message:
                      type: string
          '500':
            description: Error updating phone.

components:
  schemas:
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string

    Phone:
      type: object
      properties:
        id:
          type: string
        number:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        surname:
          type: string
        phone_id:
          type: string